optional universe name
Currently Universe::new() constructs a random name

This is particularly problematic for wasm environments which need to
configure `rand` with different features (`names` depends on `rand).
A simple workaround is to allow an optional name provided to
Universe::new(). Another benefit is that the user can
associate meaningful names with universes.