Include only needed crossbeam dependencies
When building without `par-schedule` feature one only needs the crossbeam-channel crate, not all the other parts of crossbeam. This reduces compile times when building with few features.

And when building with that feature, we only additionally need crossbeam-queue. Although in practice in that case one gets rest of crossbeam due to rayon.