Changes `Registry` to use `Arc<RwLock<S: CustomEntitySerializer>`
This way `Canon` (and other implementations of `CustomEntitySerializer`) can be used outside of
Registry, which is useful for mapping world entities to data outside of legion. A shared `Canon` should
mostly need read acces, so `RwLock` is a better choice. I don't think requiring `Sync` for `S`
is too burdensome either.