Allow use of reexported legion crate for codegen
If a crate uses legion and wants to reexport it, it works for all
functionality except the codegen.

Using the :: prefix forces the dependency to exist as a crate for calling
project - which in turn forces the user of the third-party crate to keep
legion in their own Cargo.toml and make sure the versions match over
time.